Updated tests to fit cli move change
